+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Fixed
+
+  - #3124, Fix table's media type handlers not working for all schemas - @steve-chavez
+
 ## [12.0.1] - 2023-12-12
 
 ### Fixed